F-106

+++ petcare_planner_frontend/lib/widgets/task_list_view.dart
@@ -1,11 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:petcare_planner_frontend/modal/delete_modal.dart';
+import 'package:petcare_planner_frontend/models/task.dart';
 import 'package:petcare_planner_frontend/utils/app_colors.dart';
 import 'package:petcare_planner_frontend/view_models/task_view_model.dart';
+import 'package:petcare_planner_frontend/views/task/edit_task.dart';
 import 'package:petcare_planner_frontend/widgets/task_card.dart';
 import 'package:provider/provider.dart';
 
 class TaskListView extends StatelessWidget {
-  final List tasks;
+  final List<Task> tasks;
 
   const TaskListView({super.key, required this.tasks});
 
@@ -24,21 +27,43 @@ class TaskListView extends StatelessWidget {
     }
 
     return ListView.builder(
-      padding: EdgeInsets.all(5),
+      padding: const EdgeInsets.all(5),
       itemCount: tasks.length,
       itemBuilder: (context, index) {
         final task = tasks[index];
+
         final hour = int.parse(task.time.split(':')[0]);
         final period = hour >= 12 ? 'PM' : 'AM';
 
-        return TaskCard(
-          time: task.time,
-          period: period,
-          title: task.taskTitle,
-          description: task.notes ?? task.taskType,
-          onDelete: () {
-            context.read<TaskViewModel>().deleteTask(task.id);
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => EditTaskScreen(task: task),
+              ),
+            );
           },
+          child: TaskCard(
+            time: task.time,
+            period: period,
+            title: task.taskTitle,
+            description: (task.notes != null && task.notes!.isNotEmpty)
+                ? task.notes!
+                : task.taskType,
+            onDelete: () async {
+              final bool? shouldDelete = await showDialog<bool>(
+                context: context,
+                builder: (BuildContext context) {
+                  return const DeleteModal();
+                },
+              );
+
+              if (shouldDelete == true && context.mounted) {
+                context.read<TaskViewModel>().deleteTask(task.id);
+              }
+            },
+          ),
         );
       },
     );
